+++ src/video/VideoSystem.hh
@@ -66,7 +66,7 @@ public:
 	/** Returns the current mouse pointer coordinates in the main window.
 	  * Returns std::nullopt when the main window does not have focus.
 	  */
-	[[nodiscard]] virtual std::optional<gl::ivec2> getMouseCoord() = 0;
+	[[nodiscard]] virtual std::optional<gl::vec2> getMouseCoord() = 0;
 
 	/** TODO */
 	[[nodiscard]] virtual OutputSurface* getOutputSurface() = 0;
